project structure refactored into module-wise, modified naming convention for logic models and made required changes in tests
