--- app/src/main/java/org/fossasia/openevent/app/contract/model/ICacheModel.java
@@ -1,13 +1,0 @@
-package org.fossasia.openevent.app.contract.model;
-
-public interface ICacheModel {
-
-    void saveObject(Object key, Object value);
-
-    Object getValue(Object key);
-
-    int getSize();
-
-    void clear();
-
-}
