--- app/src/main/java/org/fossasia/openevent/app/contract/presenter/IEventDetailPresenter.java
@@ -1,13 +1,0 @@
-package org.fossasia.openevent.app.contract.presenter;
-
-public interface IEventDetailPresenter {
-
-    void attach();
-
-    void detach();
-
-    void loadTickets(long eventId, boolean forceReload);
-
-    void loadAttendees(long eventId, boolean forceReload);
-
-}
