--- app/src/main/java/org/fossasia/openevent/app/contract/presenter/IEventListPresenter.java
@@ -1,15 +1,0 @@
-package org.fossasia.openevent.app.contract.presenter;
-
-public interface IEventListPresenter {
-
-    void attach();
-
-    void detach();
-
-    void loadUserEvents(boolean forceReload);
-
-    void loadOrganiser(boolean forceReload);
-
-    void logout();
-
-}
