--- app/src/main/java/org/fossasia/openevent/app/contract/presenter/ILoginPresenter.java
@@ -1,11 +1,0 @@
-package org.fossasia.openevent.app.contract.presenter;
-
-public interface ILoginPresenter {
-
-    void attach();
-
-    void detach();
-
-    void login(String email, String password);
-
-}
