--- app/src/main/java/org/fossasia/openevent/app/contract/view/ILoginView.java
@@ -1,11 +1,0 @@
-package org.fossasia.openevent.app.contract.view;
-
-public interface ILoginView {
-
-    void showProgressBar(boolean show);
-
-    void onLoginSuccess();
-
-    void onLoginError(String error);
-
-}
