+++ app/src/main/java/org/fossasia/openevent/app/data/LoginModel.java
@@ -1,29 +1,28 @@
-package org.fossasia.openevent.app.data.network;
+package org.fossasia.openevent.app.data;
 
-import org.fossasia.openevent.app.data.contract.ILoginModel;
-import org.fossasia.openevent.app.data.contract.IUtilModel;
 import org.fossasia.openevent.app.data.models.Login;
 import org.fossasia.openevent.app.data.models.LoginResponse;
 import org.fossasia.openevent.app.data.network.api.EventService;
 import org.fossasia.openevent.app.data.network.api.NetworkService;
+import org.fossasia.openevent.app.data.contract.ILoginModel;
+import org.fossasia.openevent.app.data.contract.IUtilModel;
 import org.fossasia.openevent.app.utils.Constants;
 
 import io.reactivex.Observable;
 import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.schedulers.Schedulers;
 
-public class NetworkLogin implements ILoginModel {
+public class LoginModel implements ILoginModel {
 
     private IUtilModel utilModel;
     private EventService eventService;
 
-    public NetworkLogin(IUtilModel utilModel) {
+    public LoginModel(IUtilModel utilModel) {
         this.utilModel = utilModel;
     }
 
     @Override
     public Observable<LoginResponse> login(String username, String password) {
-
         if(utilModel.isLoggedIn()) {
             return Observable.just(new LoginResponse(utilModel.getToken()));
         }
@@ -36,13 +35,14 @@ public class NetworkLogin implements ILoginModel {
             eventService = NetworkService.getEventService();
 
         return eventService
-            .login(new Login(username, password))
-            .doOnNext(loginResponse -> utilModel.saveToken(loginResponse.getAccessToken()))
-            .subscribeOn(Schedulers.io())
-            .observeOn(AndroidSchedulers.mainThread());
+                .login(new Login(username, password))
+                .doOnNext(loginResponse -> utilModel.saveToken(loginResponse.getAccessToken()))
+                .subscribeOn(Schedulers.io())
+                .observeOn(AndroidSchedulers.mainThread());
     }
 
     public void setEventService(EventService eventService) {
         this.eventService = eventService;
     }
+
 }
