+++ app/src/main/java/org/fossasia/openevent/app/data/cache/ICacheModel.java
@@ -1,0 +1,13 @@
+package org.fossasia.openevent.app.data.cache;
+
+public interface ICacheModel {
+
+    void saveObject(Object key, Object value);
+
+    Object getValue(Object key);
+
+    int getSize();
+
+    void clear();
+
+}
