+++ app/src/main/java/org/fossasia/openevent/app/data/contract/ILoginModel.java
@@ -1,0 +1,11 @@
+package org.fossasia.openevent.app.data.contract;
+
+import org.fossasia.openevent.app.data.models.LoginResponse;
+
+import io.reactivex.Observable;
+
+public interface ILoginModel {
+
+    Observable<LoginResponse> login(String username, String password);
+
+}
