+++ app/src/main/java/org/fossasia/openevent/app/events/contract/IEventsPresenter.java
@@ -1,0 +1,15 @@
+package org.fossasia.openevent.app.events.contract;
+
+public interface IEventsPresenter {
+
+    void attach();
+
+    void detach();
+
+    void loadUserEvents(boolean forceReload);
+
+    void loadOrganiser(boolean forceReload);
+
+    void logout();
+
+}
