+++ app/src/main/java/org/fossasia/openevent/app/login/contract/ILoginPresenter.java
@@ -1,0 +1,11 @@
+package org.fossasia.openevent.app.login.contract;
+
+public interface ILoginPresenter {
+
+    void attach();
+
+    void detach();
+
+    void login(String email, String password);
+
+}
