+++ app/src/main/java/org/fossasia/openevent/app/login/contract/ILoginView.java
@@ -1,0 +1,11 @@
+package org.fossasia.openevent.app.login.contract;
+
+public interface ILoginView {
+
+    void showProgressBar(boolean show);
+
+    void onLoginSuccess();
+
+    void onLoginError(String error);
+
+}
