+++ app/src/main/res/layout/activity_attendee_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.activity.AttendeeListActivity">
+    tools:context=".event.attendees.AttendeeListActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/rvAttendeeList"
