+++ app/src/main/res/layout/activity_event_details.xml
@@ -25,7 +25,7 @@
         android:paddingLeft="@dimen/activity_horizontal_margin"
         android:paddingRight="@dimen/activity_horizontal_margin"
         android:paddingTop="@dimen/activity_vertical_margin"
-        tools:context=".ui.activity.EventDetailsActivity">
+        tools:context=".event.detail.EventDetailsActivity">
 
         <TextView
             android:id="@+id/tvEventTitle"
