+++ app/src/main/res/layout/activity_login.xml
@@ -10,7 +10,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context=".ui.activity.LoginActivity">
+    tools:context=".login.LoginActivity">
 
     <android.support.design.widget.TextInputLayout
         android:layout_width="match_parent"
